chore(ops): isolate backup scope and organize HScale tests
